Add a text label rotator decorator class
Instead of peppering logic for rotating the text label in the pull down
view cell class, this breaks the rotation logic out into its own class
and then uses that to wrap the text label. Unfortunately, since the text
label property of the UITableViewCell is read only, we cannot simply use
a subclass of UILabel (hence the decoration strategy).